// types.ts
export type ViewState = 
  | 'auth'
  | 'home' 
  | 'analytics' 
  | 'teams' 
  | 'create-team'
  | 'profile' 
  | 'add-training' 
  | 'add-training-skiing'
  | 'activity-select' 
  | 'team-details' 
  | 'exercise-details' 
  | 'jump-details'
  | 'body-metrics'
  | 'all-body-metrics'
  | 'all-sessions'
  | 'coach-event-details'; // New View

export type UnitSystem = 'metric' | 'imperial';
export type Language = 'en' | 'it';
// ...
export interface BodyMetricLog {
  id: string;
  date: string;
  type: 'weight' | 'height';
  value: number; // Always stored in metric (kg or cm)
}